requirements.txt
customtkinter
scrapy
requests
python-whois
lxml
twisted
cryptography
pyopenssl
service_identity
beautifulsoup4
